+++ test/src/producer/platform_test.dart
@@ -331,11 +331,12 @@ void main() {
     );
 
     expect("""
+     import org.jetbrains.kotlin.gradle.plugin.mpp.apple.XCFramework
+    
      plugins {
           id("com.android.library")
           id("dev.buijs.klutter.gradle")
           kotlin("multiplatform")
-          kotlin("native.cocoapods")
           kotlin("plugin.serialization") version "1.6.10"
       }
       
@@ -354,20 +355,11 @@ void main() {
       }
       
       kotlin {
+          val xcf = XCFramework("Platform")
+          ios { binaries.framework { xcf.add(this) } }
+          iosSimulatorArm64 { binaries.framework { xcf.add(this) } }
           android()
-          iosX64()
-          iosArm64()
-          iosSimulatorArm64()
-      
-          cocoapods {
-              summary = "Some description for the Shared Module"
-              homepage = "Link to the Shared Module homepage"
-              ios.deploymentTarget = "14.1"
-              framework {
-                  baseName = "Platform"
-              }
-          }
-          
+                
           sourceSets {
       
               val commonMain by getting {
@@ -399,25 +391,14 @@ void main() {
                   }
               }
       
-              val iosX64Main by getting
-              val iosArm64Main by getting
-              val iosSimulatorArm64Main by getting
-              val iosMain by creating {
-                  dependsOn(commonMain)
-                  iosX64Main.dependsOn(this)
-                  iosArm64Main.dependsOn(this)
-                  iosSimulatorArm64Main.dependsOn(this)
-                  dependencies {}
+              val iosMain by getting
+              val iosSimulatorArm64Main by getting {
+                  dependsOn(iosMain)
               }
       
-              val iosX64Test by getting
-              val iosArm64Test by getting
-              val iosSimulatorArm64Test by getting
-              val iosTest by creating {
-                  dependsOn(commonTest)
-                  iosX64Test.dependsOn(this)
-                  iosArm64Test.dependsOn(this)
-                  iosSimulatorArm64Test.dependsOn(this)
+              val iosTest by getting
+              val iosSimulatorArm64Test by getting {
+                  dependsOn(iosTest)
               }
           }
       }